Create basic mentor test
